requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
Django==4.1.5
django-admin==2.0.1
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-ipware==4.0.2
django-six==1.0.4
djangorestframework==3.14.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
outcome==1.2.0
psycopg2==2.9.5
PySocks==1.7.1
python-dotenv==0.21.1
pytz==2022.7.1
screen==1.0.1
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
xlwt==1.3.0
